+++ packages/runtime/src/traits/core/Hidden.tsx
