+++ packages/runtime/src/types/index.ts
@@ -1,6 +1,6 @@
-export * from './Application';
-export * from './Component';
-export * from './Trait';
-export * from './Module';
-export * from './TraitPropertiesSchema';
-export * from './ValidResultSchema';
+export * from './application';
+export * from './component';
+export * from './trait';
+export * from './module';
+export * from './traitPropertiesSchema';
+export * from './validResultSchema';
