+++ packages/runtime/src/types/traitPropertiesSchema.ts
