+++ packages/runtime/src/utils/initStateAndMethod.ts
@@ -1,7 +1,7 @@
 import { RuntimeApplication } from '@sunmao-ui/core';
 import { Static, TSchema } from '@sinclair/typebox';
-import { Registry } from '../services/registry';
-import { StateManager } from '../services/stateStore';
+import { Registry } from '../services/Registry';
+import { StateManager } from '../services/StateManager';
 import { ModuleSchema } from '../types';
 import { parseTypeBox } from './parseTypeBox';
